+++ packages/sdk-router/src/sba/synapseBridgeAdapterModuleSet.ts
@@ -28,15 +28,21 @@ import {
   SynapseBridgeAdapterModule,
 } from './synapseBridgeAdapterModule'
 
-const SBA_BRIDGE_SUPPORTED_CHAINS = new Set<SupportedChainId>([
+const SBA_BRIDGE_ENABLED_DESTINATION_CHAINS = new Set<SupportedChainId>([
   SupportedChainId.ARBITRUM,
+  // SupportedChainId.AURORA,
   SupportedChainId.AVALANCHE,
   SupportedChainId.BASE,
+  // SupportedChainId.BLAST,
   SupportedChainId.BSC,
+  // SupportedChainId.CANTO,
+  // SupportedChainId.CRONOS,
   SupportedChainId.DFK,
   SupportedChainId.HARMONY,
   SupportedChainId.KLAYTN,
   SupportedChainId.METIS,
+  // SupportedChainId.MOONBEAM,
+  // SupportedChainId.MOONRIVER,
   SupportedChainId.OPTIMISM,
   SupportedChainId.POLYGON,
 ])
@@ -94,8 +100,7 @@ export class SynapseBridgeAdapterModuleSet extends SynapseModuleSet {
     toToken,
   }: GetBridgeTokenCandidatesParameters): Promise<BridgeTokenCandidate[]> {
     if (
-      !SBA_BRIDGE_SUPPORTED_CHAINS.has(fromChainId as SupportedChainId) ||
-      !SBA_BRIDGE_SUPPORTED_CHAINS.has(toChainId as SupportedChainId)
+      !SBA_BRIDGE_ENABLED_DESTINATION_CHAINS.has(toChainId as SupportedChainId)
     ) {
       return []
     }