+++ packages/widget-example/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.3](https://github.com/synapsecns/sanguine/compare/@synapsecns/widget-example@0.2.2...@synapsecns/widget-example@0.2.3) (2026-05-04)
+
+**Note:** Version bump only for package @synapsecns/widget-example
+
+
+
+
+
+## [0.2.2](https://github.com/synapsecns/sanguine/compare/@synapsecns/widget-example@0.2.1...@synapsecns/widget-example@0.2.2) (2026-04-28)
+
+**Note:** Version bump only for package @synapsecns/widget-example
+
+
+
+
+
 ## [0.2.1](https://github.com/synapsecns/sanguine/compare/@synapsecns/widget-example@0.2.0...@synapsecns/widget-example@0.2.1) (2026-04-01)
 
 **Note:** Version bump only for package @synapsecns/widget-example