+++ packages/widget/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.11.3](https://github.com/synapsecns/sanguine/compare/@synapsecns/widget@0.11.2...@synapsecns/widget@0.11.3) (2026-05-04)
+
+**Note:** Version bump only for package @synapsecns/widget
+
+
+
+
+
+## [0.11.2](https://github.com/synapsecns/sanguine/compare/@synapsecns/widget@0.11.1...@synapsecns/widget@0.11.2) (2026-04-28)
+
+**Note:** Version bump only for package @synapsecns/widget
+
+
+
+
+
 ## [0.11.1](https://github.com/synapsecns/sanguine/compare/@synapsecns/widget@0.11.0...@synapsecns/widget@0.11.1) (2026-04-01)